requirements.txt
pypandoc==1.5
pyspark==2.3.0
pyspark-stubs==2.3.0
pyyaml==5.3.1
py==1.9.0